Include LZMA in the project
This makes it possible to remove the liblzma dependency. However, the liblzma decompression
is around 1 second faster (for decompressing all appearance texture atlases, ~600) in debug mode.
Therefore liblzma remains for now, but it might be worth it to get rid of the liblzma dependency
in the future.

The only two functions that require liblzma are (in graphics/compression.h)
LZMA::decompressDebug
LZMA::runDebugDecompression