Fix: Tiles no longer randomly are drawn incorrectly (wrong position/texture from the atlas).
The problem was related to the alignment of the fields of Push Constants.

When drawing (not when zoomed out), random items would be drawn using incorrect textures from the
atlas and sometimes at strange positions (?).

The current fix replaces vec2 by vec4 in push constants. This appears to solve the problem, which was
probably related to GLSL alignment.

Interestingly, the problem did not occur for 64x64 items.